Modified Pintos Perl software. Split loader and kernel from now on

+++ src/threads/loader.h
@@ -37,7 +37,7 @@
 #include <stdint.h>
 
 /* Amount of physical memory, in 4 kB pages. */
-extern uint32_t init_ram_pages;
+extern uintptr_t init_ram_pages;
 #endif
 
 #endif /* threads/loader.h */
